renamed files names as per lowerCamelCase
