+++ lib/functions/curveGenerator.dart
