+++ lib/screens/otpVerificationScreen.dart
@@ -1,18 +1,18 @@
 import 'dart:ui';
 import 'package:liveasy/constants/color.dart';
 import 'package:liveasy/constants/spaces.dart';
-import 'package:liveasy/controller/hud_controller.dart';
-import 'package:liveasy/controller/timer_controller.dart';
+import 'package:liveasy/controller/hudController.dart';
+import 'package:liveasy/controller/timerController.dart';
 import 'package:liveasy/providerClass/providerData.dart';
 import 'package:flutter/material.dart';
 import 'package:liveasy/screens/navigationScreen.dart';
-import 'package:liveasy/widgets/otp_input_field.dart';
+import 'package:liveasy/widgets/otpInputField.dart';
 import 'package:modal_progress_hud_alt/modal_progress_hud_alt.dart';
 import 'package:get/get.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:liveasy/widgets/curves.dart';
-import 'package:liveasy/widgets/card_template.dart';
-import 'package:liveasy/functions/auth_functions.dart';
+import 'package:liveasy/widgets/cardTemplate.dart';
+import 'package:liveasy/functions/authFunctions.dart';
 import 'package:provider/provider.dart';
 import 'package:liveasy/constants/fontWeights.dart';
 import 'package:liveasy/constants/fontSize.dart';
